+++ src/main.rs
@@ -686,7 +686,6 @@ impl KanbanRS {
                 self.modified_since_last_saved = true;
                 self.layout_cache_needs_updating = true;
             }
-            _ => {}
         }
     }
 }